Added functionality of lives - player can lose #1
Added functionality that the script is able to keep the lives so that
when player lets enemy slip by, score is reduced by 1. So far enemies do
not shoot so the only way to lose is to let them pass by. Enemy shooting
at you will be second way that player can lose