Added Story SOunds, In Game Sounds + Story BG
Changed the story background and added sound commentary to the story +
sound effects. The Commentary i recorded myself of both of them and
modifien the alien speech. The sound effects i found and modified them a
bit. Added thunder ingame sound as well but im not yet sure if it is
there to stay.